+++ demo/config/config.exs
@@ -24,7 +24,7 @@ config :bitstyles_phoenix_demo, BitstylesPhoenixDemoWeb.Endpoint,
 
 # Configure esbuild (the version is required)
 config :esbuild,
-  version: "0.12.18",
+  version: "0.25.0",
   default: [
     args:
       ~w(js/app.js --bundle --target=es2017 --outdir=../priv/static/assets --external:/fonts/* --external:/images/*),
@@ -33,7 +33,7 @@ config :esbuild,
   ]
 
 config :dart_sass,
-  version: "1.43.4",
+  version: "1.61.0",
   default: [
     args: ~w(css/app.scss ../priv/static/assets/app.css),
     cd: Path.expand("../assets", __DIR__)