+++ demo/lib/bitstyles_phoenix_demo_web/views/error_helpers.ex
@@ -3,8 +3,6 @@ defmodule BitstylesPhoenixDemoWeb.ErrorHelpers do
   Conveniences for translating and building error messages.
   """
 
-  use Phoenix.HTML
-
   @doc """
   Translates an error message using gettext.
   """